core: sd: Moved sd particle struct into implementation
